refactor out SSE done
